statistics: split async load package
