+++ pkg/statistics/handle/storage/BUILD.bazel
@@ -22,6 +22,7 @@ go_library(
         "//pkg/sessionctx",
         "//pkg/sessionctx/variable",
         "//pkg/statistics",
+        "//pkg/statistics/asyncload",
         "//pkg/statistics/handle/cache",
         "//pkg/statistics/handle/lockstats",
         "//pkg/statistics/handle/logutil",
