+++ pkg/statistics/handle/storage/read.go
@@ -27,6 +27,7 @@ import (
 	"github.com/pingcap/tidb/pkg/parser/mysql"
 	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/statistics"
+	"github.com/pingcap/tidb/pkg/statistics/asyncload"
 	statslogutil "github.com/pingcap/tidb/pkg/statistics/handle/logutil"
 	statstypes "github.com/pingcap/tidb/pkg/statistics/handle/types"
 	"github.com/pingcap/tidb/pkg/statistics/handle/util"
@@ -539,7 +540,7 @@ func LoadHistogram(sctx sessionctx.Context, tableID int64, isIndex int, histID i
 
 // LoadNeededHistograms will load histograms for those needed columns/indices.
 func LoadNeededHistograms(sctx sessionctx.Context, statsCache statstypes.StatsCache, loadFMSketch bool) (err error) {
-	items := statistics.HistogramNeededItems.AllItems()
+	items := asyncload.AsyncLoadHistogramNeededItems.AllItems()
 	for _, item := range items {
 		if !item.IsIndex {
 			err = loadNeededColumnHistograms(sctx, statsCache, item.TableItemID, loadFMSketch, item.FullLoad)
@@ -556,12 +557,12 @@ func LoadNeededHistograms(sctx sessionctx.Context, statsCache statstypes.StatsCa
 
 // CleanFakeItemsForShowHistInFlights cleans the invalid inserted items.
 func CleanFakeItemsForShowHistInFlights(statsCache statstypes.StatsCache) int {
-	items := statistics.HistogramNeededItems.AllItems()
+	items := asyncload.AsyncLoadHistogramNeededItems.AllItems()
 	reallyNeeded := 0
 	for _, item := range items {
 		tbl, ok := statsCache.Get(item.TableID)
 		if !ok {
-			statistics.HistogramNeededItems.Delete(item.TableItemID)
+			asyncload.AsyncLoadHistogramNeededItems.Delete(item.TableItemID)
 			continue
 		}
 		loadNeeded := false
@@ -573,7 +574,7 @@ func CleanFakeItemsForShowHistInFlights(statsCache statstypes.StatsCache) int {
 			loadNeeded = loadNeeded && analyzed
 		}
 		if !loadNeeded {
-			statistics.HistogramNeededItems.Delete(item.TableItemID)
+			asyncload.AsyncLoadHistogramNeededItems.Delete(item.TableItemID)
 			continue
 		}
 		reallyNeeded++
@@ -589,13 +590,13 @@ func loadNeededColumnHistograms(sctx sessionctx.Context, statsCache statstypes.S
 	var colInfo *model.ColumnInfo
 	_, loadNeeded, analyzed := tbl.ColumnIsLoadNeeded(col.ID, true)
 	if !loadNeeded || !analyzed {
-		statistics.HistogramNeededItems.Delete(col)
+		asyncload.AsyncLoadHistogramNeededItems.Delete(col)
 		return nil
 	}
 	colInfo = tbl.ColAndIdxExistenceMap.GetCol(col.ID)
 	hg, _, statsVer, _, err := HistMetaFromStorage(sctx, &col, colInfo)
 	if hg == nil || err != nil {
-		statistics.HistogramNeededItems.Delete(col)
+		asyncload.AsyncLoadHistogramNeededItems.Delete(col)
 		return err
 	}
 	var (
@@ -649,7 +650,7 @@ func loadNeededColumnHistograms(sctx sessionctx.Context, statsCache statstypes.S
 	}
 	tbl.Columns[col.ID] = colHist
 	statsCache.UpdateStatsCache([]*statistics.Table{tbl}, nil)
-	statistics.HistogramNeededItems.Delete(col)
+	asyncload.AsyncLoadHistogramNeededItems.Delete(col)
 	if col.IsSyncLoadFailed {
 		logutil.BgLogger().Warn("Hist for column should already be loaded as sync but not found.",
 			zap.Int64("table_id", colHist.PhysicalID),
@@ -666,12 +667,12 @@ func loadNeededIndexHistograms(sctx sessionctx.Context, statsCache statstypes.St
 	}
 	_, loadNeeded := tbl.IndexIsLoadNeeded(idx.ID)
 	if !loadNeeded {
-		statistics.HistogramNeededItems.Delete(idx)
+		asyncload.AsyncLoadHistogramNeededItems.Delete(idx)
 		return nil
 	}
 	hgMeta, lastAnalyzePos, statsVer, flag, err := HistMetaFromStorage(sctx, &idx, nil)
 	if hgMeta == nil || err != nil {
-		statistics.HistogramNeededItems.Delete(idx)
+		asyncload.AsyncLoadHistogramNeededItems.Delete(idx)
 		return err
 	}
 	idxInfo := tbl.ColAndIdxExistenceMap.GetIndex(idx.ID)
@@ -713,7 +714,7 @@ func loadNeededIndexHistograms(sctx sessionctx.Context, statsCache statstypes.St
 			zap.Int64("index_id", idxHist.Info.ID),
 			zap.String("index_name", idxHist.Info.Name.O))
 	}
-	statistics.HistogramNeededItems.Delete(idx)
+	asyncload.AsyncLoadHistogramNeededItems.Delete(idx)
 	return nil
 }
